+++ src/engine/physics/PhysicsEngine.java
@@ -3,6 +3,7 @@ package engine.physics;
 
 import engine.render.Transform;
 import engine.tools.MathTools;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import java.util.ArrayList;
@@ -29,7 +30,6 @@ public class PhysicsEngine implements IPhysicsEngine {
     public void removePhysicsBody(IPhysicsBody body) {
         this.physicsBodies.remove(body);
     }
-
     public void update() {
         for(IPhysicsBody body: physicsBodies){
             body.getVelocity().addForce(Physics.GRAVITY/(body.volume()*1f), MathTools.DOWN,Force.ADD);
@@ -49,7 +49,7 @@ public class PhysicsEngine implements IPhysicsEngine {
     }
 
     private void applyVelocity(Transform transform, Velocity velocity){
-        Vector3f vel = velocity.getVelocity();
+        Vector3d vel = velocity.getVelocity();
         transform.translate(vel.x,vel.y,vel.z);
     }
